Moving reconciliation tests to controller module.
